Move qlpack functions/constants to pure
